SCRUM-25
added admin pages for beer taps work in progress
