+++ BeerTap/Components/Pages/History.razor
@@ -5,7 +5,7 @@
 @rendermode InteractiveServer
 @using BeerTap.Components.Layout
 @using BeerTap.Models
-@inject BeerTap.Services.TapHistoryService HistoryService
+@inject BeerTap.Services.TapDataService HistoryService
 @inject BeerTap.Services.UserService UserService
 
 <h3>Your Tap History</h3>
