+++ BeerTap/Models/Tap.cs
@@ -7,7 +7,9 @@ namespace BeerTap.Models
     {
         [Key]
         public Guid Id { get; set; }
+        [Required] 
         public string Name { get; set; }
+        [Required] 
         public string Type { get; set; }
         public string Topic { get; set; } = String.Empty;
         // Navigation property
